Struct zwave_controller::zwave_controller_callbacks_t

#[repr(C)]
pub struct zwave_controller_callbacks_t {

    pub on_state_updated: Option<unsafe extern "C" fn(state: zwave_network_management_state_t)>,
    pub on_error: Option<unsafe extern "C" fn(error: zwave_network_management_error_t)>,
    pub on_node_id_assigned: Option<unsafe extern "C" fn(node_id: zwave_node_id_t, included_by_us: bool)>,
    pub on_node_deleted: Option<unsafe extern "C" fn(node_id: zwave_node_id_t)>,
    pub on_node_added: Option<unsafe extern "C" fn(status: sl_status_t, nif: *const zwave_node_info_t, node_id: zwave_node_id_t, dsk: *mut u8, granted_keys: zwave_keyset_t, kex_fail_type: zwave_kex_fail_type_t, inclusion_protocol: zwave_protocol_t)>,
    pub on_new_network_entered: Option<unsafe extern "C" fn(home_id: zwave_home_id_t, node_id: zwave_node_id_t, granted_keys: zwave_keyset_t, kex_fail_type: zwave_kex_fail_type_t)>,
    pub on_keys_report: Option<unsafe extern "C" fn(csa: bool, keys: zwave_keyset_t)>,
    pub on_dsk_report: Option<unsafe extern "C" fn(input_length: u8, dsk: *mut u8, keys: zwave_keyset_t)>,
    pub on_frame_received: Option<unsafe extern "C" fn(connection_info: *const zwave_controller_connection_info_t, rx_options: *const zwave_rx_receive_options_t, frame_data: *const u8, frame_length: u16)>,
    pub on_smart_start_inclusion_request: Option<unsafe extern "C" fn(home_id: u32, already_included: bool, node_info: *const zwave_node_info_t, inclusion_protocol: zwave_protocol_t)>,
    pub on_node_information: Option<unsafe extern "C" fn(node_id: zwave_node_id_t, node_info: *const zwave_node_info_t)>,
    pub on_new_suc: Option<unsafe extern "C" fn(suc_node_id: zwave_node_id_t)>,
    pub on_node_info_req_failed: Option<unsafe extern "C" fn(node_id: zwave_node_id_t)>,
    pub on_frame_transmission_failed: Option<unsafe extern "C" fn(node_id: zwave_node_id_t)>,
    pub on_frame_transmission_success: Option<unsafe extern "C" fn(node_id: zwave_node_id_t)>,
    pub on_multicast_group_deleted: Option<unsafe extern "C" fn(group_id: zwave_multicast_group_id_t)>,
    pub on_request_neighbor_update: Option<unsafe extern "C" fn(status: u8)>,

}

@brief The following structure holds all the callbacks this module will emit.

Fields
on_state_updated: Option<unsafe extern "C" fn(state: zwave_network_management_state_t)>

This function is called when the network management state changes

on_error: Option<unsafe extern "C" fn(error: zwave_network_management_error_t)>

This function is called when an error occurs in a network management operation

on_node_id_assigned: Option<unsafe extern "C" fn(node_id: zwave_node_id_t, included_by_us: bool)>

This function is called when a node has been added to the network and has been assigned a node ID. Secure provisioning had not yet completed when this function is called @param node_id NodeID of the newly included node. @param included_by_us Set to true if we included the node, false if another controller included it.

on_node_deleted: Option<unsafe extern "C" fn(node_id: zwave_node_id_t)>

A node has been deleted from the network @param node_id NodeID of the node that was removed from our network. The value 0 can be used and indicates that a node has been excluded from a foreign network.

on_node_added: Option<unsafe extern "C" fn(status: sl_status_t, nif: *const zwave_node_info_t, node_id: zwave_node_id_t, dsk: *mut u8, granted_keys: zwave_keyset_t, kex_fail_type: zwave_kex_fail_type_t, inclusion_protocol: zwave_protocol_t)>

This function is called when a node has been assigned its security keys. This is the end of the protocol part of the inclusion process. This function is trigged both on initiated with the and_node, add_node_with_dsk add_node_proxy_inclusion. @param status Status of the provisioning @param nif_length length of the node information structure @param nif points to a node information structure @param node_id Node ID of the new node. @param dsk DSK of the new node. @param granted_keys is used to indicate which network keys were granted during bootstrapping. @param kex_fail_type This field is used to indicate which error occurred in case S2 bootstrapping was not successful

on_new_network_entered: Option<unsafe extern "C" fn(home_id: zwave_home_id_t, node_id: zwave_node_id_t, granted_keys: zwave_keyset_t, kex_fail_type: zwave_kex_fail_type_t)>

This function is called then the controller enters a new network. this can happen either after a controller set_default or after a learn mode operation.

@param home_id Home ID of the new network. @param node_id Node ID in the new network. @param granted_keys is used to indicate which network keys were granted during bootstrapping. @param kex_fail_type This field is used to indicate which error occurred in case S2 bootstrapping was not successful

on_keys_report: Option<unsafe extern "C" fn(csa: bool, keys: zwave_keyset_t)>

This function is used to inform which S2 keys have been requested during S2 bootstrapping, the receiver must call zwave_network_management_keys_set with the granted keys

@param csa This field is used to indicate if the joining node is requesting CSA @param keys This field is used to advertise the requested keys by the joining node.

on_dsk_report: Option<unsafe extern "C" fn(input_length: u8, dsk: *mut u8, keys: zwave_keyset_t)>

This function is called durring a S2 inclusion, the receiver must call zwave_network_management_dsk_set with the full dsk of the node @param input_length Number of leading bytes in the dsk which has been blanked out. @param dsk The dsk reported by the device, which input_length bytes set to zero. @param keys This field is used to advertise the requested keys by the joining node.

on_frame_received: Option<unsafe extern "C" fn(connection_info: *const zwave_controller_connection_info_t, rx_options: *const zwave_rx_receive_options_t, frame_data: *const u8, frame_length: u16)>

This callback is fired when a frame has beed received and de-encapsulated.

@param connection_info Information about the source and destination node if the frame, as well as information about how the frame was encapsulated

@param rx_options Receive specific information

@param frame_data Pointer to de-encapsulated data

@param frame_length Length of data

on_smart_start_inclusion_request: Option<unsafe extern "C" fn(home_id: u32, already_included: bool, node_info: *const zwave_node_info_t, inclusion_protocol: zwave_protocol_t)>

A SmartStart inclusion request was received.

@param home_id SmartStart DSK derived HomeId of the Smart Start node wanting to be included @param already_included Node is already included into another network.

@param node_info Node information frame of the sending node.

@param inclusion_protocol Used to indicate which protocol is intended for inclusion

on_node_information: Option<unsafe extern "C" fn(node_id: zwave_node_id_t, node_info: *const zwave_node_info_t)>

node has sent its node info either unsolicited or as a response to a ZW_RequestNodeInfo call @param node_id The source of the node information frame.

@param node_info Node information frame of the sending node.

on_new_suc: Option<unsafe extern "C" fn(suc_node_id: zwave_node_id_t)>

The SIS node Id was updated @param suc_node_id Node id of the new SUC

on_node_info_req_failed: Option<unsafe extern "C" fn(node_id: zwave_node_id_t)>

Node Info Request has failed @param node_id NodeID for which the node info request failed

on_frame_transmission_failed: Option<unsafe extern "C" fn(node_id: zwave_node_id_t)>

Frame Transmission to given node is failed @param node_id NodeID for which frame transmission failed

on_frame_transmission_success: Option<unsafe extern "C" fn(node_id: zwave_node_id_t)>

Successful Frame Transmission to given node @param node_id NodeID for which frame transmission is successful

on_multicast_group_deleted: Option<unsafe extern "C" fn(group_id: zwave_multicast_group_id_t)>

Z-Wave TX Group was deleted. @param group_id Group ID that was deleted.

on_request_neighbor_update: Option<unsafe extern "C" fn(status: u8)>

Request Node Neighbor Discovery @param status represents Neighbor Discovery Status