Struct zwave_controller::zwave_controller_transport

#[repr(C)]
pub struct zwave_controller_transport {
    pub priority: u32,
    pub command_class: zwave_command_class_t,
    pub version: u8,
    pub send_data: Option<unsafe extern "C" fn(connection: *const zwave_controller_connection_info_t, data_length: u16, data: *const u8, tx_options: *const zwave_tx_options_t, on_send_complete: on_zwave_tx_send_data_complete_t, user: *mut c_void, parent_session: zwave_tx_session_id_t) -> sl_status_t>,
    pub on_frame_received: Option<unsafe extern "C" fn(connection_info: *const zwave_controller_connection_info_t, rx_options: *const zwave_rx_receive_options_t, frame_data: *const u8, frame_length: u16) -> sl_status_t>,
}

@brief Hold information about a transport plugin.

Fields
priority: u32

@brief Priority of this plugin, transports with lowest number priority is executed last when encoding. Decoding order is made based on the Command Class identifier (first frame byte).

Recommended priorities:

1 Transport Service: TBD: Connection info should have fields with “segmentation settings” or something like this Else the transport can’t really know if this is supported by the destination

2 Security 2

3 Security 0

4 CRC 16

5 Multi Channel

command_class: zwave_command_class_t

@brief Which Z-Wave command class this transport is serving

version: u8

@brief The command class version of the command class that this transport is implementing.

send_data: Option<unsafe extern "C" fn(connection: *const zwave_controller_connection_info_t, data_length: u16, data: *const u8, tx_options: *const zwave_tx_options_t, on_send_complete: on_zwave_tx_send_data_complete_t, user: *mut c_void, parent_session: zwave_tx_session_id_t) -> sl_status_t>

@brief Send function to use when encapsulating with this transport.

@note: It is important for sender function not to pass the parent_session to the zwave_tx send_data function.

Functions registered as send_data handlers MUST return status codes according to the following:

SL_STATUS_OK indicates that the frame was accepted and encapsulated

SL_STATUS_NOT_SUPPORTED indicates that transport has nothing to encapsulate for this frame.

SL_STATUS_WOULD_OVERFLOW indicates that the frame is supported but cannot be handled, so it should not be attempted again

Any other error code (SL_STATUS_BUSY, SL_STATUS_FAIL, etc.) indicating that the frame should be encapsulated, but could not due to an error.

on_frame_received: Option<unsafe extern "C" fn(connection_info: *const zwave_controller_connection_info_t, rx_options: *const zwave_rx_receive_options_t, frame_data: *const u8, frame_length: u16) -> sl_status_t>

@brief Command handler to of the transport

Functions registered as on_frame_received transport MUST return status codes according to the following:

SL_STATUS_NOT_FOUND Indicates that the frame has no more encapsulation and must be passed manually to the upper layers by the Z-Wave Controller.

Any other code (SL_STATUS_OK, SL_STATUS_FAIL, etc.) indicating that the frame was “consumed” by the on_frame_received handler and that the handler has passed the frame to the Z-Wave Controller again if decapsulation was applied (i.e. called zwave_controller_on_frame_received())